requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
loguru==0.7.0
python-dateutil
pythorhead
requests==2.31.0
schedule==1.2.0
six==1.16.0
soupsieve==2.4.1
urllib3==2.0.3
pytz
nba_api
mastodon.py